Specify dtype more precisely

+++ python/test/unit/fem/test_mixed_mesh_dofmap.py
@@ -60,7 +60,7 @@ def test_el_dm():
     tri = coordinate_element(CellType.triangle, 1)
     quad = coordinate_element(CellType.quadrilateral, 1)
     nodes = [3 * rank + i for i in range(6)]
-    xdofs = np.array([0, 1, 4, 0, 3, 4, 1, 4, 2, 5], dtype=int) + 3 * rank
+    xdofs = np.array([0, 1, 4, 0, 3, 4, 1, 4, 2, 5], dtype=np.int32) + 3 * rank
     x = np.array(
         [[0.0, 0.0], [1.0, 0.0], [2.0, 0.0], [0.0, 1.0], [1.0, 1.0], [2.0, 1.0]], dtype=np.float64
     )
@@ -113,7 +113,7 @@ def test_el_dm_prism():
     # Create dofmaps for Geometry
     prism = coordinate_element(CellType.prism, 1)
     nodes = [0, 1, 2, 3, 4, 5]
-    xdofs = np.array([0, 1, 2, 3, 4, 5], dtype=int)
+    xdofs = np.array([0, 1, 2, 3, 4, 5], dtype=np.int32)
     x = np.array(
         [
             [0.0, 0.0, 0.0],
